OMAP4: PM: VC: allow channels use of default channel cmd_reg_addr
OMAP4's PRM_VC_CFG_CHANNEL register allows for flexibility of configuring
for various PMIC configurations. In combinations where we'd like to
use the default VC channel's cmd_reg address in a particular non-default
VC channel, we allow the use of USE_DEFAULT_CHANNEL_I2C_PARAM.

Since 0 is a valid register address, we need to increase the size of
reg storage for the flag. Depending on the PMIC and platform used,
this can be populated on the PMIC's datastructure and percolates to
VC's configuration.

+++ arch/arm/mach-omap2/vc.h
@@ -79,7 +79,7 @@ struct omap_vc_channel {
 	/* channel state */
 	u16 i2c_slave_addr;
 	u16 volt_reg_addr;
-	u8 cmd_reg_addr;
+	u16 cmd_reg_addr;
 	u8 cfg_channel;
 	u16 setup_time;
 	bool i2c_high_speed;
