+++ arch/arm/mach-omap2/voltage.h
@@ -114,7 +114,7 @@ struct omap_volt_data {
  * channel is MPU, all other domains such as IVA/CORE, could optionally
  * link their i2c reg configuration to use MPU channel's configuration if
  * required. To do this, mark in the PMIC structure's
- * i2c_slave_addr, volt_reg_addr with this macro.
+ * i2c_slave_addr, volt_reg_addr,cmd_reg_addr with this macro.
  */
 #define USE_DEFAULT_CHANNEL_I2C_PARAM  0x8000
 
@@ -143,7 +143,7 @@ struct omap_voltdm_pmic {
 	u8 vp_timeout_us;
 	u16 i2c_slave_addr;
 	u16 volt_reg_addr;
-	u8 cmd_reg_addr;
+	u16 cmd_reg_addr;
 	bool i2c_high_speed;
 	u8 i2c_mcode;
 	unsigned long (*vsel_to_uv) (const u8 vsel);
